% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinePolynomial.R
\name{.create.list.currentFactor}
\alias{.create.list.currentFactor}
\title{Create a list of all current factors (= the last ones) to be checked for all classes}
\usage{
.create.list.currentFactor(factorsClass, k)
}
\description{
Create a list of all current factors (= the last ones) to be checked for all classes
}
