% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinePolynomial.R
\name{.create.list.indexPvalue}
\alias{.create.list.indexPvalue}
\title{index of the mean p value of the current factor that we check for each class}
\usage{
.create.list.indexPvalue(currentFactorClass, model, k)
}
\description{
index of the mean p value of the current factor that we check for each class
}
