% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinePolynomial.R
\name{.create.list.pValues}
\alias{.create.list.pValues}
\title{Get the p values using the indexes}
\usage{
.create.list.pValues(mplusObject, indexPvalueClass, k)
}
\description{
Get the p values using the indexes
}
