% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{.createModelDirectory}
\alias{.createModelDirectory}
\title{Creates the specified directory if it does not exist.}
\usage{
.createModelDirectory(working_dir, model_type, overall_polynomial, classes)
}
\description{
Creates the specified directory if it does not exist.
}
