% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMplusObject.R
\name{.getAnalysis}
\alias{.getAnalysis}
\title{Creates the analysis section of an MplusObject.}
\usage{
.getAnalysis(starts, processors = parallel::detectCores())
}
\description{
Creates the analysis section of an MplusObject.
}
