% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMplusObject.R
\name{.getModel}
\alias{.getModel}
\title{Creates the model section of an MplusObject.}
\usage{
.getModel(
  usevar,
  timepoints,
  overall_polynomial,
  model_type,
  classes,
  classes_polynomial
)
}
\description{
Creates the model section of an MplusObject.
}
