% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleData.R
\name{.getSampleData}
\alias{.getSampleData}
\title{Internal function to create the testing dataset 'Diagnoses'}
\usage{
.getSampleData()
}
\description{
Internal function to create the testing dataset 'Diagnoses'
}
