% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryTable.R
\name{.getSummaryTable}
\alias{.getSummaryTable}
\title{getSummaryTable}
\usage{
.getSummaryTable(
  df,
  list_vars,
  var_labels = NA,
  group = NA,
  group.test = FALSE
)
}
\arguments{
\item{df}{A data frame containing all descriptive variables and the grouping variable}

\item{list_vars}{A vector containing the variables to summarize}

\item{var_labels}{A vector of labels for all in list_var}

\item{group}{The variable name for grouping}

\item{group.test}{Boolean, specify whether vtable will run ANOVA/ChiSq}
}
\value{
A data frame
}
\description{
getSummaryTable
}
