% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMplusObject.R
\name{.getTitle}
\alias{.getTitle}
\title{Creates the title section of an MplusObject.}
\usage{
.getTitle(classes, overall_polynomial, model_type, starts, classes_polynomial)
}
\description{
Creates the title section of an MplusObject.
}
