% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMplusObject.R
\name{.getVariable}
\alias{.getVariable}
\title{Creates the variable section of an MplusObject.}
\usage{
.getVariable(usevar, idvar, classes)
}
\description{
Creates the variable section of an MplusObject.
}
