% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMplusObject.R
\name{.splitLength}
\alias{.splitLength}
\title{Given a character vector, rebuilds each element such that the number of the 
characters and spaces between newlines is less than 50.}
\usage{
.splitLength(old_vector)
}
\description{
Given a character vector, rebuilds each element such that the number of the 
characters and spaces between newlines is less than 50.
}
