% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLCGA.R
\name{fitLCGA}
\alias{fitLCGA}
\title{fitLCGA}
\usage{
fitLCGA(
  df,
  usevar,
  timepoints,
  idvar,
  classes,
  working_dir = getwd(),
  overall_polynomial = 3,
  ref_model = NULL
)
}
\arguments{
\item{df}{A data frame containing all user variables and the ID variable}

\item{usevar}{A character vector containing variables to be used in Mplus
for analysis}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector}

\item{idvar}{A character vector containing the ID variable in the data frame}

\item{classes}{A numeric value representing the number of classes in the model}

\item{working_dir}{The directory where the results folder will be created}

\item{overall_polynomial}{A numeric value representing the polynomial order for 
the overall model (note that only linear, quadratic, and cubic models are 
supported)}

\item{ref_model}{An optional reference model to add to the list returned}
}
\value{
An MplusObject
}
\description{
fitLCGA
}
