% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset}
\alias{getDataset}
\title{getDataset}
\usage{
getDataset(model, df, idvar)
}
\arguments{
\item{model}{An MplusObject containing save data}

\item{df}{A data frame to combine with Mplus classes}

\item{idvar}{A character vector containing the ID variable in the data frame}
}
\value{
An MplusObject
}
\description{
Selects the best model from a list of MplusObjects based on a
    specified method. Current selection methods are "BIC" (select model with 
    the lowest BIC) and "BIC_LRT" (select model with the lowest BIC that also 
    has a significant LRT p-value).
}
