% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitIndices.R
\name{getFitIndices}
\alias{getFitIndices}
\title{getFitIndices}
\usage{
getFitIndices(list_models)
}
\arguments{
\item{list_models}{A list containing MplusObjects}
}
\value{
A data frame
}
\description{
Returns a selection of fit indices for a list of MplusObjects
}
