% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLongMeans.R
\name{getLongMeans}
\alias{getLongMeans}
\title{getLongMeans}
\usage{
getLongMeans(df, usevar, timepoints, group_var = NULL)
}
\arguments{
\item{df}{A data frame containing all user variables}

\item{usevar}{A character vector containing variable names}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector}

\item{group_var}{An optional grouping variable (e.g., class)}
}
\value{
A data frame
}
\description{
Wrapper around `pivot_longer` which calculate the mean for all
    user variables, then shifts them to long format
}
