% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMplusObject.R
\name{getMplusObject}
\alias{getMplusObject}
\title{getMplusObject}
\usage{
getMplusObject(
  df,
  usevar,
  timepoints,
  idvar,
  classes,
  starts,
  overall_polynomial,
  model_type
)
}
\arguments{
\item{df}{A dataframe containing all user variables and the ID variable.}

\item{usevar}{A character vector containing variables to be used in Mplus
for analysis.}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector.}

\item{idvar}{A character vector containing the ID variable in the dataframe.}

\item{classes}{A numeric value representing the number of classes in the model.}

\item{starts}{A numeric representing the number of initial stage starts for 
the model. Note that the number of final stage optimizations will be set 
as equal to half of this value.}

\item{overall_polynomial}{An numeric representing the polynomial order for 
the overall model. Note that only linear, quadratic, and cubic models are 
supported.}

\item{model_type}{A character vector representing the type of mixture model 
to create. Available options are "GBTM", "LCGA1", "LCGA2", and "LCGA3".}
}
\value{
An MplusObject
}
\description{
Provides a method to easily create an MplusObject for mixture 
    modeling. Current mixture models supported include: GBTM, with fixed 
    residual variance across time and class; LCGA1, with fixed residual 
    variance across time, but not class; LCGA2, with fixed residual variance 
    across class, but not time; and LCGA3, with unrestricted residual 
    variance across both time and class.
}
