% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpaghetti.R
\name{getSpaghetti}
\alias{getSpaghetti}
\title{Plot individual trajectories of outcome - Spaghetti plot}
\usage{
getSpaghetti(data, outvar)
}
\arguments{
\item{data}{A data frame containing all variables for the trajectory analysis.}

\item{outvar}{A character vector specifying the outcome variables at different times.}
}
\value{
A ggplot object displaying the spaghetti plot of individual trajectories.
}
\description{
Generate a spaghetti plot to visualize the individual trajectories of a given outcome across time..
}
\examples{
# Example usage:
plot <- getSpaghetti(
  data = symptoms,
  outvar = paste("sx", seq(from = 0, to = 24, by = 6), sep = "_"))

print(plot)
}
