% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryTable.R
\name{getSummaryTable}
\alias{getSummaryTable}
\title{getSummaryTable}
\usage{
getSummaryTable(df, list_vars, group_var = NULL)
}
\arguments{
\item{df}{A data frame containing all descriptive variables and the grouping variable}

\item{list_vars}{A vector containing the variables to summarize}

\item{group_var}{A factor defining the grouping variable}
}
\value{
A data frame
}
\description{
getSummaryTable
}
