% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGCM.R
\name{plotGCM}
\alias{plotGCM}
\title{plotGCM}
\usage{
plotGCM(model, timepoints, x_axis_label = "time", y_axis_label = "variable")
}
\arguments{
\item{model}{An MplusObject containing results}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector}

\item{x_axis_label}{A character vector containing text for x-axis}

\item{y_axis_label}{A character vector containing text for y-axis}
}
\description{
Plot a Growth Curve Model
}
