% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAutomatedAnalysis.R
\name{refinePolynomial}
\alias{refinePolynomial}
\title{refinePolynomial}
\usage{
refinePolynomial(model, df, usevar, idvar, working_dir = getwd())
}
\arguments{
\item{model}{The MplusObject to refine}

\item{df}{A dataframe containing all user variables and the ID variable.}

\item{usevar}{A character vector containing variables to be used in Mplus
for analysis.}

\item{idvar}{A character vector containing the ID variable in the dataframe.}

\item{working_dir}{The working directory in which the results folder will be}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector.}
}
\description{
Refines the growth factors for each class in a model
}
