% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAutomatedAnalysis.R
\name{runAutomatedAnalysis}
\alias{runAutomatedAnalysis}
\title{runAutomatedAnalysis}
\usage{
runAutomatedAnalysis(
  df,
  usevar,
  timepoints,
  idvar,
  working_dir = getwd(),
  min_k = 1,
  max_k = 6,
  max_poly = 3
)
}
\arguments{
\item{df}{A dataframe containing all user variables and the ID variable.}

\item{usevar}{A character vector containing variables to be used in Mplus
for analysis.}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector.}

\item{idvar}{A character vector containing the ID variable in the dataframe.}

\item{working_dir}{The working directory in which the results folder will be}

\item{min_k}{The minimum class structure of the model}

\item{max_k}{The maximum class structure of the model}

\item{max_poly}{The maximum polynomial order of the model}
}
\description{
Identifies the optimal model based on predefined rules using
    the BIC and LRT tests, to be used for exploratory analyses.
}
