% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGCM.R
\name{runGCM}
\alias{runGCM}
\title{runGCM}
\usage{
runGCM(df, usevar, timepoints, working_dir = getwd())
}
\arguments{
\item{df}{A data frame containing all user variables and the ID variable}

\item{usevar}{A character vector containing variables to be used in Mplus
for analysis}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector}

\item{working_dir}{The directory where the results folder will be created}
}
\value{
An MplusObject
}
\description{
Run a Growth Curve Model
}
