% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{runModel}
\usage{
runModel(
  df,
  usevar,
  timepoints,
  idvar,
  classes,
  overall_polynomial,
  model_type,
  working_dir = getwd(),
  max_starts = 4000,
  classes_polynomial = NULL
)
}
\arguments{
\item{df}{A data frame containing all user variables and the ID variable}

\item{usevar}{A character vector containing variables to be used in Mplus
for analysis}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector}

\item{idvar}{A character vector containing the ID variable in the data frame}

\item{classes}{A numeric value representing the number of classes in the model}

\item{overall_polynomial}{A numeric value representing the polynomial order for 
the overall model (note that only linear, quadratic, and cubic models are 
supported)}

\item{model_type}{A character vector representing the type of mixture model 
to create (available options are "GBTM", "LCGA1", "LCGA2", and "LCGA3")}

\item{working_dir}{The directory where the results folder will be created}

\item{max_starts}{A numeric value representing the maximum level of starting values
to use when attempting to replicate the loglikelihood before raising an 
error}

\item{classes_polynomial}{An optional character vector to pass as input to 
getMplusObject to specify the growth factors to estimate for each class}
}
\value{
An MplusObject
}
\description{
Given a set of model parameters, runs iterations of this model, 
    doubling the number of starting values until the loglikelihood value has 
    replicated at least twice within the model, and the best starting value is
    equal to that of the model with half the number of starts.
}
