% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectBestModel.R
\name{selectBestModel}
\alias{selectBestModel}
\title{selectBestModel}
\usage{
selectBestModel(list_models, method = "BIC")
}
\arguments{
\item{list_models}{A list containing MplusObjects}

\item{method}{The method to use to select the best model}
}
\value{
An MplusObject
}
\description{
Allows for the selection of the best MplusObject
}
