% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGBTM.R
\name{fitGBTM}
\alias{fitGBTM}
\title{fitGBTM}
\usage{
fitGBTM(
  df,
  usevar,
  timepoints,
  idvar,
  working_dir = getwd(),
  min_k = 1,
  max_k = 6,
  overall_polynomial = 3
)
}
\arguments{
\item{df}{A data frame containing all user variables and the ID variable}

\item{usevar}{A character vector containing variables to be used in Mplus
for analysis}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector}

\item{idvar}{A character vector containing the ID variable in the data frame}

\item{working_dir}{The directory where the results folder will be created}

\item{min_k}{The minimum class structure of the model}

\item{max_k}{The maximum class structure of the model}

\item{overall_polynomial}{A numeric value representing the polynomial order for 
the overall model (note that only linear, quadratic, and cubic models are 
supported)}
}
\value{
An MplusObject
}
\description{
Fits GBTM models from a specified minimum class to a maximum class
    using the runModel function
}
