% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAutomatedAnalysis.R
\name{fitLCGA}
\alias{fitLCGA}
\title{fitLCGA}
\usage{
fitLCGA(
  df,
  usevar,
  timepoints,
  idvar,
  k,
  working_dir = getwd(),
  overall_polynomial = 3,
  ref_model = NULL
)
}
\arguments{
\item{df}{A dataframe containing all user variables and the ID variable.}

\item{usevar}{A character vector containing variables to be used in Mplus
for analysis.}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector.}

\item{idvar}{A character vector containing the ID variable in the dataframe.}

\item{k}{The class structure of the model}

\item{working_dir}{The working directory in which the results folder will be}

\item{overall_polynomial}{The overall polynomial order of the model}

\item{ref_model}{An optional reference model to add to the list returned}
}
\description{
Fits LCGA models of specified class to test residual variance
    restrictions using the runModel function
}
