% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitIndices.R
\name{getFitIndices}
\alias{getFitIndices}
\title{getFitIndices}
\usage{
getFitIndices(
  list_models,
  type_stats = c("Title", "LL", "BIC", "Entropy", "T11_LMR_Value", "T11_LMR_PValue")
)
}
\arguments{
\item{list_models}{A list containing MplusObjects}

\item{type_stats}{A character vector containing the statistics to return}
}
\value{
A data frame
}
\description{
Returns a selection of fit indices for a list of MplusObjects
}
