% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitIndices.R
\name{getFitIndices}
\alias{getFitIndices}
\title{getFitIndices}
\usage{
getFitIndices(list_models)
}
\arguments{
\item{list_models}{A list of MplusObjects}
}
\value{
A summary of models in list as df
}
\description{
Returns the fit indices for a list of MplusObjects
}
