% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMplusObject.R
\name{getMplusObject}
\alias{getMplusObject}
\title{getMplusObject}
\usage{
getMplusObject(
  df,
  usevar,
  timepoints,
  idvar,
  classes,
  starts,
  overall_polynomial,
  model_type,
  classes_polynomial = NULL
)
}
\arguments{
\item{df}{A data frame containing all user variables and the ID variable}

\item{usevar}{A character vector containing variables to be used in Mplus
for analysis}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector}

\item{idvar}{A character vector containing the ID variable in the data frame}

\item{classes}{A numeric value representing the number of classes in the model}

\item{starts}{A numeric value representing the number of initial stage starts for 
the model (note that the number of final stage optimizations will be set 
as equal to half of this value)}

\item{overall_polynomial}{A numeric value representing the polynomial order for 
the overall model (note that only linear, quadratic, and cubic models are 
supported)}

\item{model_type}{A character vector representing the type of mixture model 
to create (available options are "GBTM", "LCGA1", "LCGA2", and "LCGA3")}

\item{classes_polynomial}{An optional character vector to pass as input to 
getMplusObject to specify the growth factors to estimate for each class}
}
\value{
An MplusObject
}
\description{
Provides a method to easily create an MplusObject for mixture 
    modeling. Current mixture models supported include: GBTM, with fixed 
    residual variance across time and class; LCGA1, with fixed residual 
    variance across time, but not class; LCGA2, with fixed residual variance 
    across class, but not time; and LCGA3, with unrestricted residual 
    variance across both time and class.
}
