% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGMobject.R
\name{LGMobject}
\alias{LGMobject}
\title{Create Mplus model objects for Latent Growth Modelling (LGM)}
\usage{
LGMobject(
  data,
  outvar,
  catvar = FALSE,
  idvar,
  k,
  starting_val,
  estimator = c("MLR", "ML", "WLSMV", "WLS"),
  transformation = c("LOGIT", "PROBIT"),
  lgm_type = c("gcm", "gbtm", "lcga", "gmm"),
  residuals = c("fixed", "free_time", "free_class", "relaxed"),
  random_effect = c("class_invariant", "class_variant"),
  polynomial = 1,
  timescores,
  timescores_indiv = FALSE,
  mplus_model = NULL,
  output,
  plot,
  save
)
}
\arguments{
\item{data}{A data frame containing all variables for the trajectory analysis.}

\item{outvar}{A character vector specifying the outcome variables at different times.}

\item{catvar}{A logical value indicating whether the outcome variable is categorical. Default is \code{FALSE}.}

\item{idvar}{A character string specifying the ID variable.}

\item{k}{An integer specifying the number of latent classes for the model.}

\item{starting_val}{A numeric value specifying the number of random starting values to generate for the initial optimization stage.
Note that the number of final stage optimizations will be set as equal to half of this value.}

\item{estimator}{A character string to specify the estimator to use in the analysis. Default is 'MLR'.}

\item{transformation}{A character string to specify the latent response variable transformation to use when the outcome variable is categorical. Default is \code{LOGIT}.}

\item{lgm_type}{A character string specifying the type of latent growth model to build. Options include: gcm, gbtm, lcga and gmm.}

\item{residuals}{A character string specifying the structure of the residual variance (variability of the errors at each time point).
Options include:
\itemize{
\item "fixed" (fixed residual variance across time and class),
\item "free_time" (free residual variance across time),
\item "free_class" (free residual variance across class),
\item "relaxed" (relaxed residual variance across both time and class).
}}

\item{random_effect}{A character string specifying the structure of the random effects.
Options include:
\itemize{
\item "class_invariant" – Random effect variances are constrained to be equal across latent classes.
\item "class_variant" – Random effect variances are freely estimated for each latent class.
}}

\item{polynomial}{An integer specifying the order of the polynomial used to model trajectories. Supported values are:
1 (linear), 2 (quadratic), 3 (cubic). Default is 1.}

\item{timescores}{A numeric vector specifying the time scores for the model. If \code{timescores_indiv = TRUE},
a character vector should be used to specify variables with individually varying times of observation.}

\item{timescores_indiv}{A logical value indicating whether to use individually varying times of observation for the outcome variable. Default is \code{FALSE}.}

\item{mplus_model}{A character string specifying a custom model using Mplus language. Default is \code{NULL}.}

\item{output}{A character vector specifying the requested Mplus output options for the model.}

\item{plot}{A character string specifying the requested Mplus plot options for the model.}

\item{save}{A character string specifying the type of results to be saved by Mplus.}
}
\value{
A list of class \code{mplusObject} with elements specifying sections of an Mplus input file for conducting latent growth modeling.
}
\description{
Facilitates the specification of Mplus LGM objects, allowing control over the number of latent classes, polynomial growth terms, residual variance structures, and inclusion of random effects.
Support Growth Curve Models (GCM), Growth-Based Trajectory Models (GBTM), Latent Class Growth Analysis (LCGA) and Growth Mixture Models (GMM).
Once specified, the model object can be estimated using the \code{runLGM()} function.
}
\details{
The \code{LGMobject} function facilitates and automates the appropriate model specification for conducting latent growth modeling in Mplus.
It creates the relevant sections of an Mplus input file, including: TITLE, VARIABLE, ANALYSIS, MODEL, OUTPUT, PLOT, and SAVEDATA.

This function builds upon the capabilities of the \code{\link[MplusAutomation]{mplusObject}}function
from the MplusAutomation package.
}
\examples{
# Example usage:
GBTM_object <- LGMobject(
  data = symptoms,
  outvar = paste("sx", seq(from = 0, to = 24, by = 6), sep = "_"),
  idvar = "id",
  catvar = FALSE,
  k = 3L,
  starting_val = 500,
  lgm_type = "gmm",
  residuals = "relaxed",
  random_effect = "class_invariant",
  polynomial = 3,
  timescores = seq(from = 0, to = 24, by = 6),
  timescores_indiv = FALSE,
  output = c("TECH1", "TECH14", "SAMPSTAT", "STANDARDIZED"),
  plot = "PLOT3",
  save = "FSCORES"
)
}
\seealso{
\code{\link[MplusAutomation]{mplusObject}} for creating an mplusObject.
\code{\link{runLGM}} for conducting latent growth modelling with an mplusObject.
}
