% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModel.R
\name{plotModel}
\alias{plotModel}
\title{plotModel}
\usage{
plotModel(mplus_model, figure_caption = "default")
}
\arguments{
\item{mplus_model}{MplusObject containing results}

\item{figure_caption}{Character vector containing text to be added as a caption
to the ggplot object}
}
\value{
A ggplot object
}
\description{
Plots class trajectories of an MplusObject
}
