% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{runModel}
\usage{
runModel(
  df,
  usevar,
  timepoints,
  idvar,
  classes,
  overall_polynomial,
  model_type,
  working_dir = getwd(),
  max_starts = 4000
)
}
\arguments{
\item{df}{A dataframe containing all user variables and the ID variable.}

\item{usevar}{A character vector containing variables to be used in Mplus
for analysis.}

\item{timepoints}{A vector containing the timepoints corresponding 
to the elements in the usevar vector.}

\item{idvar}{A character vector containing the ID variable in the dataframe.}

\item{classes}{A numeric value representing the number of classes in the model.}

\item{overall_polynomial}{An numeric representing the polynomial order for 
the overall model. Note that only linear, quadratic, and cubic models are 
supported.}

\item{model_type}{A character vector representing the type of mixture model 
to create. Available options are "GBTM", "LCGA1", "LCGA2", and "LCGA3".}

\item{working_dir}{The working directory in which the results folder will be}

\item{max_starts}{A numeric representing the maximum level of starting values
to use when attempting to replicate the logliklihood before raising an 
error.}
}
\value{
An MplusObject containing results
}
\description{
Provides a method to run a model with a given class structure
    and polynomial order, testing that the logliklihood value has replicated
    at least twice within the model, as well as in the model with half the
    number of starts.
}
