% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectBestModel.R
\name{selectBestModel}
\alias{selectBestModel}
\title{selectBestModel}
\usage{
selectBestModel(list_models, selection_method = "BIC_LRT")
}
\arguments{
\item{list_models}{A list containing MplusObjects}

\item{selection_method}{A character vector representing the method to use for
model selection}
}
\value{
An MplusObject
}
\description{
Selects the best model from a list of MplusObjects based on a
    specified method. Current selection methods are "BIC" (select model with 
    the lowest BIC) and "BIC_LRT" (select model with the lowest BIC that also 
    has a significant LRT p-value).
}
